+++ innative/validate.cpp
@@ -123,7 +123,7 @@ void innative::AppendError(const Environment& env, ValidationError*& errors, Mod
   err->error[len] = 0;
   err->code       = code;
   err->m          = -1;
-  if(m >= env.modules && (m - env.modules) < env.n_modules)
+  if(m >= env.modules && size_t(m - env.modules) < env.n_modules)
     err->m = m - env.modules;
 
   do
